% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcess.r
\name{convToBinary}
\alias{convToBinary}
\title{Convert specified columns in a dataset to binary 0 or 1.}
\usage{
convToBinary(df, colNms)
}
\arguments{
\item{df}{data frame}

\item{vifCutoff}{VIF cutoff to use (default = 4)}
}
\description{
Convert specified columns in a dataset to binary 0 or 1.
}

