% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preProcess.r
\name{convToDummy}
\alias{convToDummy}
\title{Dummy code the specified columns in a dataset.  The returned
columns will be of the form colNm1_val1, colNm1_val2, etc}
\usage{
convToDummy(dfOrig, colNms)
}
\arguments{
\item{colNms}{a list of column names to dummy code}

\item{df}{data.frame or data.table}
}
\description{
Dummy code the specified columns in a dataset.  The returned
columns will be of the form colNm1_val1, colNm1_val2, etc
}

