% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_wald.R
\name{.bracket_non_an_chr}
\alias{.bracket_non_an_chr}
\title{Place double square brackets around non-alphanumeric characters}
\usage{
.bracket_non_an_chr(x)
}
\arguments{
\item{x}{character vector.}
}
\value{
Character, with non-alphanumeric characters now being bracketed by double square brackets.
}
\description{
Useful for string comparisons
}
\examples{

}
