% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wald_stats.R
\name{.get_var_ind}
\alias{.get_var_ind}
\title{Get indices of variables to include in wald test}
\usage{
.get_var_ind(est, vcov, var, match_condn, match_print)
}
\value{
A numeric vector o
}
\description{
Get indices of variables to include in wald test
}
