% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wald_stats.R
\name{.get_wald_stats}
\alias{.get_wald_stats}
\title{Calculate wald statistic from coef estimates and vcov-mat}
\usage{
.get_wald_stats(est, vcov, var, match_condn, match_print, intercept_nm)
}
\description{
Calculate wald statistic from coef estimates and vcov-mat
}
