% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_wald.R
\name{.test_wald}
\alias{.test_wald}
\title{Calculate wald statistic from coef estimates and vcov-mat}
\usage{
.test_wald(est, vcov, var, match_condn, match_print, intercept_nm)
}
\description{
Calculate wald statistic from coef estimates and vcov-mat
}
