% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_wald.R
\name{.test_wald_ind}
\alias{.test_wald_ind}
\title{Calculate Wald statistic given summary statistics}
\usage{
.test_wald_ind(est, vcov, ind, var)
}
\arguments{
\item{est}{numeric vector. Estimates of parameters.}

\item{vcov}{numeric matrix. Variance-covarance matrix of parameters.}

\item{ind}{logical or numeric vector. Indicates which are the parameters
in \code{est} that we need to test against the null hypothesis of being zero.}
}
\value{
A tibble with the following elements:
\tabular{ll}{
stat \tab Wald test statistic \cr
df \tab Degrees of freedom \cr
p \tab P-value \cr
}
}
\description{
Calculate Wald statistic given summary statistics
}
\examples{
est_vec <- c(0.2,5,1)
vcov_mat <- as.matrix(rbind(c(1, 0.3, 0.1), c(0.3, 4, 0.4), c(0.1, 0.4, 2.2)))
ind_vec_lgl <- c(TRUE, FALSE, TRUE)
ind_vec_num <- c(1, 3)
# using logical vector to indicate which parameters are to be tested
.test_wald(est = est_vec, vcov = vcov_mat, ind = ind_vec_lgl)
# using numeric vector to indicate which parameters are to be tested
.test_wald(est = est_vec, vcov = vcov_mat, ind = ind_vec_num)
}
