% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mod_fml_as_chr.R
\name{get_mod_fml_as_chr}
\alias{get_mod_fml_as_chr}
\title{Create model formula as character from components}
\usage{
get_mod_fml_as_chr(
  var_dep,
  var_exp = NULL,
  var_exp_spline = NULL,
  var_re = NULL,
  var_offset = NULL,
  var_conf = NULL,
  rhs_text = NULL,
  int = TRUE,
  sub_exp = 0
)
}
\arguments{
\item{var_dep}{character. Name of dependent variable. Goes on left-hand side of formula.}

\item{var_exp}{character vector. Specifies explanatory variables to be included as is in model.}

\item{var_exp_spline}{named list of named list(s).. For each element, the name denotes then name of the explanatory variable
for which a spline term is required. Each element is a named list, where the following names denote the following things.
The name \code{"splines_pkg_fn"} specifies the spline function to be used from the \code{splines} pkg, e.g.
\code{splines_pkg_fn = "ns"} means that the \code{ns} function from the \code{splines} pkg is used. The
\code{knots}-named element specifies the knots to be used. No other named elements are supported at present
(but can be easily added).}

\item{var_re}{character vector. Specifies individual random effects in the \code{lme4} notation..
If named, then the name of each
element specifies the variable that has a random effect for each level of the corresponding element. If unnamed,
then it is assumed that the intercept has a random effect. For example, \code{var_re = c("time" = "x1",
"x2")} adds \code{(time|x1) + (1|x2)} to the formula. If blank, then no random effect terms are added.}

\item{var_offset}{character vector. Text to be added as \code{"offset(<var_offset>)"},
e.g. \code{var_offset = "log(days)"} adds \code{offset(log(days))} to the model. If blank,
then no offset term is added. That that this can be a character vector, but it's probably
the case that \code{lme4::{g}lmer}, at least, requires that there is only one offset term.}

\item{var_conf}{character vector. Specifies confounders to be added.}

\item{rhs_text}{character. Text simply to be concatenated to the model formula. Note that "+"
is added before adding this.}

\item{int}{logical. If \code{TRUE}, then formula includes an intercept. Default is \code{TRUE}.}

\item{sub_exp}{non-negative integer. Specifies maximum number of explanatory variables
to be excluded from the model formulae. If for example \code{sub_exp = 1}, then
model formula as strings are returned where each explanatory variable is excluded on its own (but
all other explanatory variables of interest and other terms (such as confounders and random effects))
are kept. If greater than one, then all possible combinations of explanatory variables to exclude
are excluded, along with all possible combinations of explanatory variables to exclude of size less than this.}
}
\value{
A named list, where the name specifies the type of formula and the
corresponding element is the model formula as a stringr. The full model formula (with all explanatory variables)
has name \code{'full'}. The null model (with zero explanatory variables of interest but
including all random effects, offsets, confounders and any additional text specified by \code{rhs_text}) has name
null. The full model less one or more explanatory variables is denoted by the name of the excluded
explanatory variables, separated by \code{~p~} (the 'p' denoted plus).
}
\description{
Create model formula as character from components
}
\details{
This was written for the following reasons:
- So that if we want to fit multiple, similar models (for example, the null model, the full model
and all models less one variable) the formulae could be obtained quickly.
- So that if we want to link model formula to character arguments in functions, it can be done easily.
This is useful if, for example, the same function that fits the model is also used to create
plots of model output, as the key argument (the variables for which we want plots) can be
used to specify the model formula and then be used to create the plots.
}
