% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{get_pred}
\alias{get_pred}
\title{Extract predicted values}
\usage{
get_pred(mod)
}
\arguments{
\item{mod}{model object of class 'lm', 'lmerMod' or 'glmerMod'.}
}
\value{
A tibble with column \code{pred}./
}
\description{
Extract predicted values
}
