% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{get_resid}
\alias{get_resid}
\title{Extract residuals}
\usage{
get_resid(mod)
}
\arguments{
\item{mod}{object of class 'lm'.}
}
\value{
A tibble with an entry for each observation for each of the following columns:
\tabular{rr}{
.resid_raw \tab raw residuals (predicted - actual) \cr
.resid_std \tab standardised residuals ((predicted - actual)(standard deviation of response)) \cr
}
}
\description{
Extract residuals
}
\details{
Notes:
broom::augment.lm gives .resid_std column as standardised deviance residuals (standardised by standard error calculated using all samples)
deviance is good as it's more N(0,1) (when standardised, I would guess) than Pearson (can't remember what that is now, maybe the working?)
}
