% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{get_se}
\alias{get_se}
\title{Extract the standard error of prediction}
\usage{
get_se(fit, data)
}
\arguments{
\item{mod}{model object of class 'lm', 'lmerMod' or 'glmerMod'.}
}
\description{
Extract the standard error of prediction
}
