% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wald_stats.R
\name{get_wald_stats}
\alias{get_wald_stats}
\title{Calculate Wald statistics}
\usage{
get_wald_stats(
  fit,
  var,
  match_condn = "start",
  match_print = "inexact",
  intercept_nm = "(Intercept)"
)
}
\arguments{
\item{fit}{object of class 'glmerMod' or 'list'. Methods are defined for each of these
to then extract the estimates and the var-cov matrix before calculating the Wald statistics.
If a list, then must have elements with names \code{est} and \code{vcov}, corresponding
to the numeric vector of estimates and the variance-covariance matrix of}

\item{var}{character vector. Name(s) of variables for which the Wald test is required. If \code{"0"},
then all variables are automatically included. If \code{"-1"}, then all variables except
intercept are included. No default.}

\item{match_condn}{'exact', 'start' or 'any'. If \code{'exact'}, then
variables in \code{var} match coefficient names in \code{fit} only if they match exactly (e.g.
\code{'abc'} matches \code{'abc'} but neither \code{'a'} nor \code{'c'} do). If \code{ 'start'},
then they match only if they correspond to the start of coefficient names in \code{fit} (e.g.
\code{'abc'} and \code{'a'} match \code{'abc'} \code{'c'} does not). If \code{'any'},
then they match as long as they appear at some point in the coefficient name in \code{fit} (e.g.
\code{'abc'}, \code{'a'} and \code{'c'} all match \code{'abc'}). Default is \code{'start'}.}

\item{match_print}{'all', 'inexact' or 'none'. If \code{'all'}, then
all matched coefficient names in \code{fit} are printed. If \code{'inexact'}, then
only inexact matches are printed. If \code{'none'}, then none are printed. Default is \code{'all'}.}

\item{intercept_nm}{character. Name for intercept coefficient. Default is \code{"(Intercept)"}. Only
used if \code{class(fit) = 'list'}.}
}
\value{
A tibble with the following elements:
\tabular{ll}{
test \tab 'wald' \cr
stat \tab test statistic \cr
df \tab degrees of freedom \cr
p \tab p-value \cr
}
}
\description{
Calculate Wald statistics
}
