% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelEval.r
\name{getBalancedAcc}
\alias{getBalancedAcc}
\title{Get's the balanced accuracy by choosing a probability
threshold value where Specificity == Sensitivity at the resolution
of the crit parameter.}
\usage{
getBalancedAcc(prob, obs, posLabel, negLabel, crit = 1e-04, verbose = F)
}
\arguments{
\item{prob}{predicted probability}

\item{obs}{observed labels}

\item{posLabel}{the positive label (e.g. YES)}

\item{negLabel}{the negative label (e.g. NO)}

\item{crit}{criical value resolution for the difference in sensitivity==specificity}

\item{verbose}{boolean flag to print out the difference value}
}
\description{
Get's the balanced accuracy by choosing a probability
threshold value where Specificity == Sensitivity at the resolution
of the crit parameter.
}

