% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelEval.r
\name{getBinEvalVals}
\alias{getBinEvalVals}
\title{Pulls the AUC and Gini values from predicted probability and
observed labels.}
\usage{
getBinEvalVals(prob, obs)
}
\arguments{
\item{prob}{predicted probability}

\item{obs}{observed labels}
}
\description{
Pulls the AUC and Gini values from predicted probability and
observed labels.
}

