% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelEval.r
\name{getCMVals}
\alias{getCMVals}
\title{Pulls confusion matrix metrics from predicted and observed labels,
works for both binomial and multinomial.  Calls caret confusionMatrix()
and then returns as a vector.}
\usage{
getCMVals(pred, obs)
}
\arguments{
\item{pred}{vector of predicted labels}

\item{obs}{vector of observed labels}
}
\description{
Pulls confusion matrix metrics from predicted and observed labels,
works for both binomial and multinomial.  Calls caret confusionMatrix()
and then returns as a vector.
}

