% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelEval.r
\name{getGINI}
\alias{getGINI}
\title{Compute the Gini coefficient}
\usage{
getGINI(actual, predicted)
}
\arguments{
\item{actual}{binary vector}

\item{predicted}{real-valued vector that defines the ranking}
}
\description{
This function computes the Gini coefficient as (2*(AUC-.5))
}

