% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preProcess.r
\name{getSummaryTable}
\alias{getSummaryTable}
\title{Get's the basic summary stats for a data set.}
\usage{
getSummaryTable(df)
}
\arguments{
\item{df}{data.frame or data.table}
}
\description{
Get's the basic summary stats for a data set.
}

