% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{ll}
\alias{ll}
\title{Extract log-likelihoods}
\usage{
ll(mod, get_ll = NULL)
}
\arguments{
\item{mod}{All objects that have a method for them in \code{broom::glance} or
\code{broom.mixed::glance} or \code{modutils:::.;;} or \code{stats::logLik}.
Run \code{library(broom.mixed); utils::methods(broom:glance)}
or \code{library(stats); utils::methods(stats::logLik)} to see a list.}

\item{get_ll}{function. A function that takes a model as input and returns
the log-likelihood. Provided so that the user need not write a method for
a package if the log-likelihood function is not available.}
}
\value{
Numeric vector of length one containing the log-likelihood.
}
\description{
A wrapper around \code{logLik} essentially. Implemented
just in case there is a model that does not have a \code{logLik} generic.
}
\examples{
set.seed(3)
test_tbl <- data.frame(x = 1:10)
test_tbl$y <- 2 * test_tbl$x + rnorm(10)
mod_lm <- lm(y ~ x, data = test_tbl)
ll(mod_lm)
}
