% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{make_path}
\alias{make_path}
\title{Create some folders for output}
\usage{
make_path(path, replc = FALSE)
}
\arguments{
\item{path}{the filepath for folder to create}

\item{replc}{logical for whether to replace the folder}
}
\description{
Create some folders for output
}

