% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelEval.r
\name{multiClassMetricsMeans}
\alias{multiClassMetricsMeans}
\title{Get's overall evaluation metrics for a multiclass classification model.
Includes overall accuracy and kappa, and then the arithmetic mean of 
all classes for sensitivty, specificity, detection rate, and balanced accuracy.
Additionally includes one-vs-all calculations of AUC and log loss, which is
presented as the mean value across all one-vs-all calculations.  
 
Designed to work as caret summaryFunction where data is a data.frame
with columns 'pred' and 'obs' for predicted and observed labels, and the class
specific probabilities (columns as label names).}
\usage{
multiClassMetricsMeans(data, lev = NULL, model = NULL)
}
\arguments{
\item{data}{data.frame with columns 'pred', 'obs' for predicted and observed labels, and a column for each class probability.}

\item{lev}{caret summaryFunction optional parameter}

\item{model}{caret summaryFunction optional parameter}
}
\description{
Adapted from code by Zach Meyer: http://www.r-bloggers.com/error-metrics-for-multi-class-problems-in-r-beyond-accuracy-and-kappa/
}

