% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelEval.r
\name{multiClassSummary}
\alias{multiClassSummary}
\title{Get's evaluation metrics for a multiclass classification model.
Includes calculation of a custom balanced accuracy that is the product
of a specific labels balanced accuracy (mean of spec and sens) times
the minimum of spec or sens.}
\usage{
multiClassSummary(data, lev = NULL, model = NULL)
}
\arguments{
\item{data}{data.frame with columns 'pred' and 'obs' for predicted and observed labels}

\item{lev}{caret summaryFunction optional parameter}

\item{model}{caret summaryFunction optional parameter}
}
\description{
Designed to work as caret summaryFunction where data is a data.frame
with columns 'pred' and 'obs' for predicted and observed labels.
}

