% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit_mod.R
\name{refit_mod}
\alias{refit_mod}
\title{Fit a model using "easier" optimisation settings}
\usage{
refit_mod(
  mod_expr_as_chr,
  pkg,
  family,
  theta_fix = NULL,
  env = rlang::caller_env()
)
}
\arguments{
\item{mod_expr_as_chr}{character. Expression to run model as a character,
wrapped in a try call, e.g. \code{try(lm(y~x, data = data_mod))}.}

\item{pkg}{'lme4'. Package used to fit model.}

\item{family}{'nb'. 'nb' denotes negative binomial.}

\item{theta_fix}{numeric. If \code{NULL}, then negative binomial model is
refit using less stringent \code{tolPwrss} parameter (changing it from 1e-7 to 1e-6,
and then to 1e-5 if that doesn't fit).}

\item{env}{environment. Environment in which to fit model. Default is calling environment
of \code{refit_mod}.}
}
\description{
Intended to automatically fit a model using easier settings.
}
\details{
For mixed-effects negative binomial models fit using the `lme4` package, if \code{is.null(theta_fix)},
then model is refit using less stringent \code{tolPwrss} parameter (changing it from 1e-7 to 1e-6,
and then to 1e-5 if that doesn't fit). If provided, then model is refit using \code{theta_fix} as
the fixed dispersion parameter.
}
\examples{
test_tbl <- data.frame(x = (1:10)^2, grp = rep(1:5, each = 2), dur = 1:10)
test_tbl$y <- (test_tbl$x + test_tbl$grp * 2)
suppressMessages(library(lme4))
mod_fml_as_chr <- modutils::get_mod_fml_as_chr(var_dep = "y", var_exp = "x", var_offset = "dur", var_re = "grp")[2]

mod_expr_as_chr <- paste0("try(lme4::glmer.nb(", mod_fml_as_chr, ", data = test_tbl))")

# refitting nb model with more lenient tolPwrss optimisation argument
mod_refit_tolpwrss <- suppressWarnings(refit_mod(mod_expr_as_chr, pkg = 'lme4', family = 'nb'))

# refit with fixed theta
theta <- lme4::getME(mod_orig, "glmer.nb.theta")

mod_refit_theta_fix <- suppressWarnings(refit_mod(mod_expr_as_chr, pkg = 'lme4', family = 'nb', theta_fix = theta/2))
lme4::getME(mod_refit_theta_fix, "glmer.nb.theta"))
}
