% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcess.r
\name{removeZeroVar}
\alias{removeZeroVar}
\title{This is rip of caret function: just looks at % of most frequent, not 
relative to the second most frequent}
\usage{
removeZeroVar(df, freqCutoff = 0.95, colsToKeep = NULL)
}
\arguments{
\item{df}{data frame}

\item{freqCutoff}{minimum % of the most common value (simple tweak of caret function)}

\item{colsToKeep}{vector of column names that should not be removed}
}
\description{
This is rip of caret function: just looks at % of most frequent, not 
relative to the second most frequent
}

