% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_lr.R
\name{test_lr}
\alias{test_lr}
\title{Perform likelihood ratio test}
\usage{
test_lr(mod_large, mod_small, var = NULL)
}
\arguments{
\item{mod_large}{Model in which other model(s) are nested.}

\item{mod_small}{list or model. If a model, then coerced to a
list. Names of list elements are taken to be the names of the
variables missing in \code{mod_small} that are included in \code{mod_large}.}

\item{var}{character vector. If provided, then it
gives the names of the variables omitted in \code{mod_small}. If omitted,
then the names of the variables missing in \code{mod_small} are automatically
determined. Default is \code{NULL}.}
}
\value{
A tibble with the following columns:
\tabular{rrrrr}{
var \tab names of variables missing in mod small \cr
test \tab "lr" \cr
stat \tab likelihood ratio \cr
df \tab degrees of freedom \cr
p \tab p-value \cr
}
}
\description{
Perform likelihood ratio test
}
\examples{
 set.seed(2106); n_id <- 30
test_tbl <- tibble::tibble(id = rep(as.character(1:n_id), each = 2),
                          grp = rep(as.character(1:2), each = n_id),
                          y_re = rep(rnorm(n_id), each = 2),
                          x = rnorm(n_id * 2),
                          y = x^4 * ifelse(grp == "1", 3, 0) + y_re + rnorm(20, sd = 0.3))
test_tbl$y <- test_tbl$y +  abs(min(test_tbl$y)) + 0.001
test_tbl$y <- test_tbl$y/max(test_tbl$y + 0.001)

mod_lm <- lm(y ~ splines::ns(x, 3), data = test_tbl)
mod_lm_int <- lm(y ~ 1, data = test_tbl)
test_lr(mod_lm, mod_lm_int, 'x')
}
