% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcess.r
\name{balanceClasses}
\alias{balanceClasses}
\title{Simple balancing of classes for multinomial. If up-sample, do with 
replacement, if down sample no replacement.}
\usage{
balanceClasses(df, target, classSize)
}
\arguments{
\item{df}{data frame}

\item{target}{target name of the dataset}

\item{classSize}{the number of observations in each target class}
}
\description{
Simple balancing of classes for multinomial. If up-sample, do with 
replacement, if down sample no replacement.
}

