% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelEval.r
\name{getBalancedProbCut}
\alias{getBalancedProbCut}
\title{Get's the probability cutoff for a binomial classification 
model that minimizes the difference between sensitivity and
specificity.}
\usage{
getBalancedProbCut(prob, obs, posLabel, negLabel, probSeq = seq(0.001, 0.999,
  by = 0.001))
}
\arguments{
\item{prob}{predicted probability}

\item{obs}{observed labels}
}
\description{
Get's the probability cutoff for a binomial classification 
model that minimizes the difference between sensitivity and
specificity.
}

