% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modBuild.r
\name{getCVFolds}
\alias{getCVFolds}
\title{Simple fold index generator, all idx used,
randomly sorted.}
\usage{
getCVFolds(n, K)
}
\arguments{
\item{n}{dataset size}

\item{K}{number of folds}
}
\value{
vector of folds as integers
}
\description{
Simple fold index generator, all idx used,
randomly sorted.
}

