% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelEval.r
\name{multiClassSummary}
\alias{multiClassSummary}
\title{Use multiClassMetrics(data, lev, model).  Left here for backwards compat.}
\usage{
multiClassSummary(data, lev = NULL, model = NULL)
}
\arguments{
\item{data}{data.frame with columns 'pred' and 'obs' for predicted and observed labels}

\item{lev}{caret summaryFunction optional parameter}

\item{model}{caret summaryFunction optional parameter}
}
\description{
Use multiClassMetrics(data, lev, model).  Left here for backwards compat.
}

